[SSHD-518] Add support for SFTP 'copy-file' extension

+++ sshd-core/src/main/java/org/apache/sshd/client/subsystem/sftp/SftpClient.java
@@ -90,7 +90,7 @@ public interface SftpClient extends SubsystemClient {
     public static class Attributes {
         public final Set<Attribute> flags = EnumSet.noneOf(Attribute.class);
         public long size;
-        public byte type;
+        public int type;
         public int uid;
         public int gid;
         public int perms;
