+++ sshd-core/src/test/java/org/apache/sshd/WindowTest.java
@@ -105,7 +105,7 @@ public class WindowTest extends BaseTestSupport {
                             return new ServerUserAuthService(session) {
                                 @SuppressWarnings("synthetic-access")
                                 @Override
-                                public void process(byte cmd, Buffer buffer) throws Exception {
+                                public void process(int cmd, Buffer buffer) throws Exception {
                                     authLatch.await();
                                     super.process(cmd, buffer);
                                 }
